pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-coffee-project"
version = "0.1.0"
description = "Professional deep learning pipeline for coffee bean image classification."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Deep Coffee Project Team" }]
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "scikit-learn>=1.4.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "jupyter>=1.0.0",
  "tqdm>=4.66.0",
  "Pillow>=10.0.0",
  "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]
future = [
  "gradio>=4.0.0",
  "streamlit>=1.30.0",
  "huggingface_hub>=0.20.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["configs*", "src*", "tests*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
